Fix segfault

+++ statbasic.h
@@ -2,7 +2,7 @@
 #define STATBASIC_H
 
 void diff(float *before, float *after, int len);
-void sort(float data[], int len);
+void sort(int data[], int len);
 float sum(float *data, int len);
 float dataSquare(float *data, int len);
 float sumOfDegrees(float ySquare, float sum, int len);
